+++ src/backend/marsha/core/lti/__init__.py
@@ -1,13 +1,12 @@
 """LTI module that supports LTI 1.0."""
 import re
-from urllib.parse import unquote, urlparse
+from urllib.parse import urlparse
 
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.http.request import validate_host
 from django.utils.datastructures import MultiValueDictKeyError
 
-from oauthlib import oauth1
 from pylti.common import LTIException, verify_request_common
 
 from ..models import ConsumerSite
@@ -216,38 +215,6 @@ class LTI:
             "course_name": self.context_title,
             "course_run": None,
         }
-
-    def sign_post_request(self, url, lti_parameters):
-        """Sign a request by adding all oauth parameters."""
-        passport = self.get_passport()
-
-        client = oauth1.Client(
-            client_key=passport.oauth_consumer_key, client_secret=passport.shared_secret
-        )
-        # Compute Authorization header which looks like:
-        # Authorization: OAuth oauth_nonce="80966668944732164491378916897",
-        # oauth_timestamp="1378916897", oauth_version="1.0", oauth_signature_method="HMAC-SHA1",
-        # oauth_consumer_key="", oauth_signature="frVp4JuvT1mVXlxktiAUjQ7%2F1cw%3D"
-        _uri, headers, _body = client.sign(
-            url,
-            http_method="POST",
-            body=lti_parameters,
-            headers={"Content-Type": "application/x-www-form-urlencoded"},
-        )
-
-        # Parse headers to pass to template as part of context:
-        oauth_dict = dict(
-            param.strip().replace('"', "").split("=")
-            for param in headers["Authorization"].split(",")
-        )
-
-        signature = oauth_dict["oauth_signature"]
-        oauth_dict["oauth_signature"] = unquote(signature)
-        oauth_dict["oauth_nonce"] = oauth_dict.pop("OAuth oauth_nonce")
-
-        lti_parameters.update(oauth_dict)
-
-        return lti_parameters
 
     @property
     def resource_link_title(self):